backend/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
python-dotenv
djangorestframework-simplejwt
requests
django-ratelimit